Move "set record file" to the record function from the settings function - Delete from settings - Add to record - Add new checkbox option to bypass the action if the GSD is currently recording.   (This allows a single button that first sets the date-time then toggles recording.    with the option enable it will only change the filename before starting the video)
