+++ src/functions/record/actionId.ts
@@ -2,4 +2,5 @@ export enum ActionId {
 	Record = 'record',
 	RecordStatus = 'recordStatus',
 	RecordTime = 'recordTime',
+	RecordFileName = 'recordFileName',
 }
