+++ src/functions/record/actions.ts
@@ -1,5 +1,5 @@
 import { ActionId } from './actionId'
-import { getOptNumber } from '../../util'
+import { getOptNumber, getOptString } from '../../util'
 import { ReqType } from '../../enums'
 import { sendCommand } from '../../connection'
 import type { CompanionActionDefinitions } from '@companion-module/base'
@@ -30,6 +30,33 @@ export function create(_model: GoStreamModel, state: RecordStateT): CompanionAct
 					newState = state.isRecording === true ? 0 : 1
 				}
 				await sendCommand(ActionId.Record, ReqType.Set, [newState])
+			},
+		},
+		[ActionId.RecordFileName]: {
+			name: 'Set Record FileName',
+			options: [
+				{
+					type: 'textinput',
+					label: 'FileName',
+					id: 'RecordFileName',
+					required: true,
+					default: '',
+				},
+				{
+					type: 'checkbox',
+					label: 'Update ONLY IF not recording',
+					id: 'SafeMode',
+					default: true,
+				},
+			],
+			callback: async (action, context) => {
+				if (action.options.SafeMode && state.isRecording) return
+				// else
+
+				const rawString = getOptString(action, 'RecordFileName')
+				const newName = await context.parseVariablesInString(rawString)
+				// allow but replace ":" with a valid character, so user can specify system time in the variable
+				await sendCommand(ActionId.RecordFileName, ReqType.Set, [newName.replaceAll(':', '_')])
 			},
 		},
 	}
